requirements.txt
Django~=5.1.5
djangorestframework~=3.15.2
django-cors-headers~=4.6.0
python-dotenv~=1.0.1
selenium~=4.27.1
psycopg2-binary~=2.9.9
pgvector~=0.3.6